requirements.txt
# Core PDF processing
pdfplumber>=0.10.0
pypdf>=4.0.0

# OCR (optionnel - pour PDFs scannés)
pytesseract>=0.3.10
pdf2image>=1.16.0

# Image processing (optionnel - pour pré-traitement OCR)
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Data processing
pandas>=2.0.0

# JSON validation
jsonschema>=4.20.0

# Configuration
pyyaml>=6.0

# Reports
tabulate>=0.9.0
